moved sparse and sparsegeometry to _core
