+++ src/sisl/io/siesta/_help.py
@@ -5,8 +5,8 @@ import numpy as np
 
 import sisl._array as _a
 from sisl import SislError
+from sisl._core.sparse import _rows_and_cols
 from sisl.messages import warn
-from sisl.sparse import _rows_and_cols
 
 try:
     from . import _siesta
