+++ src/sisl/physics/sparse.py
@@ -8,10 +8,10 @@ from scipy.sparse import SparseEfficiencyWarning, csr_matrix
 
 import sisl._array as _a
 import sisl.linalg as lin
+from sisl._core.sparse import issparse
+from sisl._core.sparse_geometry import SparseOrbital
 from sisl._internal import set_module
 from sisl.messages import warn
-from sisl.sparse import issparse
-from sisl.sparse_geometry import SparseOrbital
 
 from ._matrix_ddk import matrix_ddk, matrix_ddk_nc, matrix_ddk_nc_diag, matrix_ddk_so
 from ._matrix_dk import matrix_dk, matrix_dk_nc, matrix_dk_nc_diag, matrix_dk_so
